adicionando logica de salvar anotações no planner

+++ database/migrations/2025_06_25_205525_add_user_id_to_communities_table.php
@@ -8,17 +8,21 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::table('communities', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->after('id'); // adiciona depois do id
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('communities', 'user_id')) {
+            Schema::table('communities', function (Blueprint $table) {
+                $table->unsignedBigInteger('user_id')->after('id');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
     {
         Schema::table('communities', function (Blueprint $table) {
-            $table->dropForeign(['user_id']);
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('communities', 'user_id')) {
+                $table->dropForeign(['user_id']);
+                $table->dropColumn('user_id');
+            }
         });
     }
 };
